Dont crash if there is an unknown error at login. Remove a bug that do x86_64 arch fail to connect to server : long is 8 bytes on that arch, use int for 4 bytes integer.

+++ src/net/messageout.h
@@ -50,7 +50,7 @@ class MessageOut
 
         void writeByte(char value);          /**< Writes a byte. */
         void writeShort(short value);        /**< Writes a short. */
-        void writeLong(long value);          /**< Writes a long. */
+        void writeLong(int value);          /**< Writes a long. */
 
         /**
          * Writes a string. If a fixed length is not given (-1), it is stored
